load/requirements.txt
requests
google-cloud-storage
google-cloud-bigquery
flask
gunicorn
werkzeug